images/search-api/requirements.txt
aioredis==1.3.1
fastapi==0.63.0
elasticsearch[async]==7.10.1
pydantic==1.7.3
orjson==3.4.7
uvicorn==0.13.3
elasticsearch-dsl==7.3.0
pyjwt[crypto]==2.1.0
backoff==1.11.1